Implement the full Machine hierarchy and MainApp
