+++ README.md
@@ -1,9 +1,9 @@
 # jv-oop
 
-- Create class Machine containing methods `public abstract void doWork()` and `public abstract void stopWork()`.  
-- Create classes `Truck`, `Bulldozer`, and `Excavator` that will inherit from `Machine`.
+- Create class core.basesyntax.Machine containing methods `public abstract void doWork()` and `public abstract void stopWork()`.  
+- Create classes `core.basesyntax.Truck`, `core.basesyntax.Bulldozer`, and `core.basesyntax.Excavator` that will inherit from `core.basesyntax.Machine`.
 - In those classes override `doWork()`, so it will print a message that a certain machine started its work.
 - Override `stopWork()` as well. It should print messages that certain machines stopped working.
-- In the `MainApp` class create a `Machine` array with `Truck`, `Bulldozer`, and `Excavator` and call methods `doWork()` and `stopWork()` in a loop.
+- In the `MainApp` class create a `core.basesyntax.Machine` array with `core.basesyntax.Truck`, `core.basesyntax.Bulldozer`, and `core.basesyntax.Excavator` and call methods `doWork()` and `stopWork()` in a loop.
 
 #### [Try to avoid these common mistakes, while solving task](./checklist.md)
