--- src/main/java/Bulldozer.java
@@ -1,11 +1,0 @@
-public class Bulldozer extends Machine {
-    @Override
-    public void doWork() {
-        System.out.println("Bulldozer started its work");
-    }
-
-    @Override
-    public void stopWork() {
-        System.out.println("Bulldozer stopped working");
-    }
-}
