+++ src/main/java/core/basesyntax/Bulldozer.java
@@ -1,0 +1,13 @@
+package core.basesyntax;
+
+public class Bulldozer extends Machine {
+    @Override
+    public void doWork() {
+        System.out.println("core.basesyntax.Bulldozer started its work");
+    }
+
+    @Override
+    public void stopWork() {
+        System.out.println("core.basesyntax.Bulldozer stopped working");
+    }
+}
