+++ src/main/java/core/basesyntax/Machine.java
@@ -1,3 +1,5 @@
+package core.basesyntax;
+
 public abstract class Machine {
     public abstract void doWork();
     public abstract void stopWork();
