+++ src/main/java/core/basesyntax/MainApp.java
@@ -1,5 +1,21 @@
 package core.basesyntax;
 
 public class MainApp {
+    public static void main(String[] args) {
+        Machine[] machines =  new Machine[3];
+
+        machines[0] = new Truck();
+        machines[1] = new Bulldozer();
+        machines[2] = new Excavator();
+
+        for (Machine machine : machines) {
+            machine.doWork();
+        }
+        System.out.println();
+
+        for (Machine machine : machines) {
+            machine.stopWork();
+        }
+    }
 
 }
